dropdown, helpdrawer, idletimeout, steplist, table updated to sb7 syntax
